Change codegen to use load/stores everywhere
This is likely less efficient. But the previously used method of codegen was simply invalid. The llvm ir has an ordering which should be respected, at least in case of atomics. It may also contain function calls which do not directly see any use, but have side effects. It simply won't work to see the llvm ir as being solely tree-like, we must respect it having a temporal ordering